+++ template-parts/content-blog.php
@@ -76,24 +76,7 @@ if ( 'grid' === $blog_layout ) {
                         
                         <?php if ( $show_comments ) : ?>
                             <span class="comments-link">
-                                <?php
-                                $comment_count = get_comments_number();
-                                if ( $comment_count == 0 ) {
-                                    $comments_text = __( '0 Comments', 'mec_theme' );
-                                    $screen_reader_text = __( '0 comments', 'mec_theme' );
-                                } elseif ( $comment_count == 1 ) {
-                                    $comments_text = __( '1 Comment', 'mec_theme' );
-                                    $screen_reader_text = __( '1 comment', 'mec_theme' );
-                                } else {
-                                    /* translators: %s: number of comments */
-                                    $comments_text = sprintf( __( '%s Comments', 'mec_theme' ), number_format_i18n( $comment_count ) );
-                                    $screen_reader_text = sprintf( __( '%s comments', 'mec_theme' ), number_format_i18n( $comment_count ) );
-                                }
-                                ?>
-                                <a href="<?php comments_link(); ?>">
-                                    <span aria-hidden="true"><?php echo esc_html( $comments_text ); ?></span>
-                                    <span class="screen-reader-text"><?php echo esc_html( $screen_reader_text ); ?></span>
-                                </a>
+                                <?php mec_theme_comments_link_markup(); ?>
                             </span>
                         <?php endif; ?>